+++ NEWS.md
@@ -1,4 +1,19 @@
 
+# rquery 1.2.1 2018/12/16
+
+ * versions without _nse() suffix.
+ * parse ^ power.
+ * reflexive arguments to non-sql nodes.
+ * Add .[]-variable decleration notation.
+ * Narrow suggests.
+ * update docs.
+ * add order_expr.
+ * declare : as an inline operator.
+ * allow concatinating compatible pipelines.
+ * check for table name consistency.
+ * NULL/NA parsing.
+ * Force order limit to be >= 0.
+
 # rquery 1.2.0 2018/11/06
 
  * fix name lookup issue.